functions/trainer/requirements.txt
functions-framework
google-cloud-storage
google-cloud-firestore

# Machine Learning y Datos
pandas
numpy
scikit-learn
xgboost